% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{densityEst}
\alias{densityEst}
\title{Density estimation using logspline approach}
\usage{
densityEst(y, x, ymargin, lbound, ubound, K, nymargin = 1000, log = TRUE)
}
\arguments{
\item{y}{a time series of data}

\item{x}{a sequence of numbers from 1 to n where n is the number of time series}

\item{ymargin}{data interval in the y-direction}

\item{lbound}{lower bound for the support for the density}

\item{ubound}{upper bound for the support for the density}

\item{K}{number of initial number of knots}

\item{nymargin}{desired length of the seqence}

\item{log}{Does the logarithmic transformation apply to data? Default is TRUE.}
}
\description{
It nonparametrically estimates a time series of density functions using logspline approach considering the time ordering of the densities.
}
\examples{
l <- 20
n <- 2000
z <- seq(1000,5000,length.out=n)
mean <- seq(mean(z)-400,mean(z)+200,length.out=l)
sd <- seq(200,400,length.out=l)
y <- lapply(1:l, function(X) rnorm(z,mean[X],sd[X]))
x <- c(1:l)
densityEst(y,x)
}
