% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plots.R
\name{plots}
\alias{plots}
\title{Plotting density estimation using logspline approach}
\usage{
plots(final)
}
\arguments{
\item{final}{A list of outputs returns from densityEst function}
}
\description{
It nonparametrically estimates a time series of density functions using logspline approach considering the time ordering of the densities.
}
